feat: Support nested filters in MockGenerator
